ALL HAIL THE MIGHTY WEEKVIEW CONTROL
Fall down and pray, pray for your pathetic lives in the face of the one true control, THE WEEKVIEW!
Also, appointments and busies need to be stored in blocks of 30 minutes (even if their recorded start times are not). And having a 'is beginning' bool values would be helpful as well.